requirements.txt
Django==3.0.7
djangorestframework==3.11.0
python-dotenv==0.13.0
django-cors-headers==3.4.0
